django/requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.4
django-admin-sortable2==0.7.1
django-ckeditor==5.7.1
django-cors-headers==3.0.2
django-extensions==2.2.1
django-filter==2.2.0
django-js-asset==1.2.2
django-minio-storage==0.2.2
django-rest-swagger==2.2.0
django-solo==1.1.3
djangorestframework==3.10.2
future==0.17.1
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
minio==4.0.21
openapi-codec==1.3.2
Pillow==2.1.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
shortuuid==0.5.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
uWSGI==2.0.18
mysqlclient==1.4.4